Completed the Autocomplte.
The Autocomplete loads an instance of VariableScanner, which loads a
list of Snippets from disk, as well as scans the sketch for vaiable and
class names. As soon as a word is 3 characters long the autocomplete
looks at all Snippets and creates a Popupmenu that lists all possible
fill-in possibilities.